requirements.txt
yfinance~=0.2.28
pandas~=2.0.3
numpy~=1.26.0
sqlalchemy~=2.0.0
psycopg2-binary~=2.9.9
python-dotenv~=1.0.0